Update LOAD_PATH so Puppet can find parent providers.
Puppet does not make parent providers avaible to Ruby's LOAD_PATH
when they are contained in modules that are plugin-synced to an agent.

I opened this ticket #15074 to explain the issue in Puppet.

This code updates all of the providers to append their lib
directory to Puppet's LOAD_PATH so that these modules will work
with pluginsync.