removed colocation sort, order is important
Example:

   a
  / \
 b   c

colocation a b
colocation a c

this tells resource a to colocate with b and c, if b and c are on diffrent nodes,
a will randomly choose a node,
b and c will not care about colocation

colocation b a
colocation c a

this tells the resources b and c to colocate with a,
a will not care about colocation, but b and c will go to the node where resource a is.

The result is completly different depending on the order.