(MODULES-1503) Fix parameter quotes
The cs_primitive providers have a :parameters property where the value
of a parameter could be a list of values. They therefore must be quoted
so that the whole list is associated with the key of the parameter
during the flush operation. This patch adds the quotes and updates the
unit tests for them.