Stop ignoring order of primitives in colocation constraints
cs_colocation, when used with the pcs provider, makes the incorrect
assumption that the order of primitives in colocation constraints is
arbitrary. It is not, and the following must not be equivalent:

cs_colocation { 'vip_with_service':
  primitives => [ 'nginx_service', 'nginx_vip' ],
}

cs_colocation { 'vip_with_service':
  primitives => [ 'nginx_vip', 'nginx_service' ],
}

These mean different things. The first example ensures that nginx_vip
runs on whatever node nginx_service runs on, and if nginx_service
can't be started on any node, Pacemaker won't even attempt to start
nginx_vip. However, if nginx_vip cannot be started on any node,
nginx_service will continue to be available.

The second example means the opposite. In it, nginx_service is
configured to run wherever nginx_vip runs, and if nginx_vip can't run
anywhere, nginx_service won't start. But if nginx_service cannot be
started, nginx_vip is still available.

Note that the provider's expected order is directly reversed compared
to the one that standalone Pacemaker would use (no matter whether pcs
or crm is used). This is intentional as per commit 8e218d9.

At any rate, for the pcs provider the order of resources is wrong, and
injecting the sort to mask that issue is nonsense. As a result, drop
artificial sorting for the primitives, fix the order, and use them
exactly as specified.

Fixes issue puppet-community/puppet-corosync#150.