Allow 1-element arrays for primitives in cs_group
Having only one primitive in a group is a perfectly valid use
case. For example, one might have constraints related to an arbitrary
number of primitives created with create_resources, and the simplest
way of doing that is to place them all of them in a group, and have
the constraint point to that. Disallowing the possibility of ending up
with just one primitive in the group does not make much sense.

Update the check to allow 1-element arrays. Also, remove a misleading
comment that also looks like it was copied from some other code that
talked about colocations, not groups.

Fixes issue puppet-community/puppet-corosync#152.