Fix acceptance tests for RHEL6 and Ubuntu 14.04
Acceptance tests were failing on Ubuntu Trusty because the quorum
members must include at least one node that matches the bind address,
and were failing on CentOS 6 because the corosync package on CentOS 6
does not properly stop the pacemaker service before trying to restart
the corosync service. This patch fixes the tests by fixing the
quorum_members array and adding an exec to stop pacemaker when
refreshing corosync.