Log the update being loaded by crm.
Sometimes updates fail to apply and it's not clear why. I'm not
sure how to get the output (stderr?) out of the crm exec, but this
change is a step nearer to finding out what's going on. By logging
the update being run it's fairly easy to replicate it on the command
line and see the error message.