Fix detached keystone node epoch mismatch
Pacemaker maintains an internal database, which is used for
configuration storage. Each update of this database increases a counter,
called "epoch", which should have the same value cluster-wide.

If an update operation comes to a previously detached node, a conflict
will occur. Pacemaker does not allow updating this database on a node,
which epoch value is lower than the epoch value of a cluster leader.

However, it is safe to force the update during the bootstrap phase. It
can be implemented by directly manipulating the values by using
"cibadmin" utility (and providing the delta XML) instead of "crm".