Fix variable type to allow autkey to be a string
This is related to PR #69 which introduced the capability of corosync
class to take `authkey` as a string if `authkey_source` is set to
'string'.

Problem appeared at commit 42e13c76bc29027f6406199b9ab260c009157f8e when
switching to Puppet 4 data types: `$authkey` was assigned a
Stdlib::AbsolutePath data type.