member section is deprecated in favor of nodelist
With recent release of corosync, it is recommended to migrate config
file to nodelist because member section is deprecated.

bindnetaddr might also conflicts with nodelist, so we remove bindnetaddr
if nodelist exists.

We rely on the value of set_votequorum to remove member section and
bindnetaddr from corosync.conf. This parameter is tied to quorum_members
in the corosync class.

spec: swap bindnetaddr and ringnumber in corosync.conf

Fix #421 and #422