Puppet Strings Reference & Corosync Crypto
Updated each of the existing manifest files to generate a REFERENCE.md
via puppet strings.

Corosync Crypto Update

Replaced 'secauth' in corosync.conf with the current crypto_hash and
crypto_cipher parameters and updated the main class to take the appropriate
arguments.

Note that this change is not used on older versions of
Ubuntu, RedHat/CentOS, and Debian which are still supported by this
module but utilize Corosync 1.x instead of 2.x. Those OSs which are
still supported continue to use the 'secauth' parameter and ignore the
valuse of crypto_hash and crypto_cipher.

Additionally, the documentation has been fleshed out to detail creating and
deploying a secure key for Corosync's communication based on the corosync-keygen
process. This also includes an updated to deploy the key file in a binary format
as the application does not specify which key formats are supported.