corosync-qdevice Bug Correction
This primarily addreses the case where a new cluster with a quorum
device is deployed instead of a partially configured cluster. The
primary problem is that the pcs command to add the quorum device won't
work if the quorum config is present in corosync.conf so we need to
ensure that all these steps happen before that is attempted.

Additionally the following changes are included:
* pcs auth nolonger depends on the contents of corosync.conf to find the
node list. Instead, all nodes are authorized by name.
* corosync-qdevice now requires a cluster_name. This may not be a true
requriemnet, however, for brand new clusters a temporary cluster must be
created to avoid circular dependencies on corosync.conf and that
requires a name to generate via pcs.
* The addition of the qdevice now uses the config status to check
whether it needs to run.