enable_packages compatibility
The puppetlabs-stdlib library includes a function called `ensure_packages` which
accepts a list of packages and ensures they are present without generating
catalog errors. This simplifies concurrent deployment of modules which depend on
the same packages, however, it only works when the packages are declared with
exactly the same values.

If a declaration exisits with different resource parameters (including a different value
for ensure) the ensure_packages function will attempt to redefine the existing
resource and generate a catalog error. As ensure_packages always sets the ensure
property to 'present' no other value will ever allow the function to work as
expected.

Relevant changelog entry:
https://github.com/puppetlabs/puppetlabs-stdlib/blob/c4243084b449a5c801d99c0d5ea4c27abb09b1c6/CHANGELOG.md