Unroll qdevice Hiera Data
The original implementation of the qdevice class sourced package names for
pcs and corosync_qnetd from hiera. It doesn't make sense to introduce a change
in methodology in an unrelated PR, particularly when the defaults are always the
same, so this switches those values to be parameter level defaults.

Additionally, includes minor corrections to the puppet strings and updates the
corresponding REFERENCE file as spurious types were included in declarations.