Reduce delay for systemd stop of pacemaker.service
The sane-y default of a 30 minute timeout is too much when travis kills
hanging executions after ten minutes.

The real cause for the restart to hang is not very clear yet (seems to
be related to stonith being enabled while not having anything
configured), but this at least makes the tests finish.