Don't fail when set_votequorum=true and cluster_name is set
Corosync allows setting cluster_name so the user does not have to
manually define and specify multicast addresses (the addresses are
derived from the cluster name).

This means that Puppet catalog compilation should not fail() when
set_votequorum is true and cluster_name is set but quorum_members and
multicast_address are empty. Only fail when neither quorum_members, nor
multicast_address nor cluster_name are defined.

This change updates the logic and extends the spec tests to cover
the described case.