feat: include name in nodelist
When the member list is names, not IP addresses, and the member name
array is empty use the former to populate the name: in the nodelist.

This addresses a parsing issue in `pcs cluster config` described in the
redhat article below. It also ensures other commands that depend on the
config file work as expected.

https://access.redhat.com/solutions/6992919