replace `String#concat` with `Array#push` to avoid frozen string violation
with Ruby 3.0 strings are frozen by default causing the catalog compiler
to throw errors when one of the native corosync module types is used.

this change replaces the use of the string concatenation operator
with an array that gets joined into a string for further use.

the functionality and data passed to external systems should remain
unchanged, making this change backwards compatible.